Refactor Create/Update/Upsert directives to DRY
